+++ libs/fsm/tests/event_queue.c
@@ -45,7 +45,6 @@
 #include "onesize/onesize.h"
 #include "slist/slist.h"
 #include "strlib/strlib.h"
-#include "pal/pal.h"
 #include "fsm/fsm.h"
 
 #define AM_EVT_A AM_EVT_USER
@@ -144,11 +143,7 @@ static enum am_rc fsmq_b(struct am_fsmq *me, const struct am_event *event) {
 }
 
 int main(void) {
-    struct am_event_state_cfg cfg = {
-        .crit_enter = am_pal_crit_enter,
-        .crit_exit = am_pal_crit_exit,
-    };
-    am_event_state_ctor(&cfg);
+    am_event_state_ctor(/*cfg=*/NULL);
 
     {
         static char pool[1 * AM_POOL_BLOCK_SIZEOF(struct am_event)] AM_ALIGNED(