+++ libs/hsm/tests/event_queue.c
@@ -61,7 +61,6 @@
 #include "onesize/onesize.h"
 #include "slist/slist.h"
 #include "strlib/strlib.h"
-#include "pal/pal.h"
 #include "hsm/hsm.h"
 
 #define AM_EVT_A AM_EVT_USER
@@ -159,10 +158,7 @@ static enum am_rc hsmq_s2(struct am_hsmq *me, const struct am_event *event) {
 }
 
 int main(void) {
-    struct am_event_state_cfg cfg = {
-        .crit_enter = am_pal_crit_enter, .crit_exit = am_pal_crit_exit
-    };
-    am_event_state_ctor(&cfg);
+    am_event_state_ctor(/*cfg=*/NULL);
 
     {
         static char pool[1 * AM_POOL_BLOCK_SIZEOF(struct am_event)] AM_ALIGNED(