+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [0-based versioning](https://0ver.org/).
 - Prune `AM_ENABLE_WARNING()` and `AM_DISABLE_WARNING()` macros.
 - Fix AM_GET_MACRO_2_() documentation.
 - Relocate compiler independent macros to `common/macros.h`.
+- Allow configuration set to NULL in `am_event_state_ctor()`.
 
 ### Added